Hook up window split component to store
