+++ browser/src/UI/Shell/ShellView.tsx
@@ -4,6 +4,8 @@
 
 import * as React from "react"
 
+import { Provider } from "react-redux"
+
 import * as Platform from "./../../Platform"
 
 import { getKeyEventToVimKey } from "./../../Input/Keyboard"
@@ -47,7 +49,9 @@ export class ShellView extends React.PureComponent<IShellViewComponentProps, {}>
                             </div>
                             <div className="container full">
                                 <div className="stack">
-                                    <WindowSplits windowManager={this.props.windowManager} />
+                                    <Provider store={this.props.windowManager.store}>
+                                        <WindowSplits />
+                                    </Provider>
                                 </div>
                                 <Overlays />
                             </div>
